requirements.txt
google-generativeai
scikit-learn
numpy
pydantic==2.6.1
tiktoken==0.6.0
requests
openai
jieba
anthropic
rich==13.7.1
pandas
openpyxl
rank_bm25
questionary
pyfiglet
prompt_toolkit
pydantic-settings==2.2.1
faiss-cpu==1.11.0
langchain-community==0.0.30
langchain-core==0.1.37
langchain==0.1.13
langchain-text-splitters==0.0.1
unstructured==0.12.6
